+++ src/forging_blocks/presentation/errors/error_status_code_mapper.py
@@ -2,7 +2,7 @@
 
 from dataclasses import replace
 
-from forging_blocks.presentation.error_view_model import ErrorViewModel
+from forging_blocks.presentation.errors.error_view_model import ErrorViewModel
 
 
 class ErrorStatusCodeMapper:
